Version: 2008-06-08 ksh93t-
Sourced from ast-open-history version 2008-06-08

+++ src/lib/libsum/sum-lmd.c
@@ -42,7 +42,7 @@ typedef struct Lmd_s
 	_SUM_LMD_
 	struct
 	{
-	unsigned int	context;
+	uintmax_t	context;
 	}		context;
 } Lmd_t;
